BACKEND/requirements.txt
asgiref==3.11.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
Django==6.0
django-cors-headers==4.3.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
google-auth==2.41.1
google-auth-oauthlib==1.2.3
gspread==6.2.1
idna==3.11
oauthlib==3.3.1
pillow==12.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
sqlparse==0.5.4
tzdata==2025.2
urllib3==2.6.0
